+another Terje-alike variant, this time with number split to three parts.
It turned out that on KBL all base100 variants run at approximately
the same speed and that variant with Terje-style fix point is not the fastest
among them (although differences are near noice level).
I thought that may be that the reason could be a latency chains, so I made
another variant, where the bulk of calculations runs as 3 independent chains.
But the measurements show that it does not matter.
The bottleneck is elswhere, most likely in instruction issue or retirement.
So, the best strategy for further improvements shell be to try to reduce
the number of instructions.