more tweaks of 1mul
Make it more like Terje's, with multiplication by 5 and shifting
of mask to theright.
It takes more instructions but improves latency.
And, unexpectedly for me, throughput improves too. Probably, the latency is so high that even
such OoO monsters, like modern Intel cores can't cover all of it.

Now, the question is - why the latency is so high?
I expected that LEA/AND/LEA = 3 clocks per step, but measurements show
something like 4-4.5 clocks per step.