+bin2ascii_2chains_base100 - same ideas as bin2ascii_1chain_base100, but lower and upper digits are converted as separate chains
On IvyB this variant runs with the lowest latency (esp. compiled by gcc)
and the throghput is very close to the best.
Compilers do all multiplications by 25 with pair of LEAs, so there are
only 2 mul instructions.
The variant looks pretty good except one major flow - it is not written in portable C.
It depends on gcc extension __int128.